feat: 회원의 모든 반려동물별 전체 산책 통계 조회 기능 추가

+++ src/main/java/com/concord/petmily/domain/walk/dto/WalkStatisticsDto.java
@@ -15,5 +15,5 @@ public class WalkStatisticsDto {
     private double totalDistanceKm; // 총 산책 거리(KM)
     private long totalDurationMinutes; // 총 산책 시간(Minutes)
     private double averageDistanceKm; // 평균 산책 거리(KM)
-    private double averageDurationMinutes; // 평균 산책 시간(Minutes)
+    private long averageDurationMinutes; // 평균 산책 시간(Minutes)
 }
